WebApi degli articoli FUNGE :+1:

+++ ArticoliWebService/Models/Articoli.cs
@@ -16,12 +16,13 @@ namespace ArticoliWebService.Models
         public string Um { get; set; }
         public string CodStat { get; set; }
         [Range(0,100, ErrorMessage = "I pezzi devono essere compresi tra 0 e 100")]
-        public int? PzCart {get; set;}
+        public Int16? PzCart {get; set;}
         public int? IdIva { get; set;}
         public int? IdFamAss { get; set;}
         [Range(0.01, 100, ErrorMessage = "I pezzi devono essere compresi tra 0.01 e 100")]
         public double? PesoNetto { get; set;}
         public DateTime? DataCreazione {get; set;}
+        
         public virtual ICollection<Ean> Barcode {get; set;}
         public virtual Ingredienti Ingredienti {get; set;}
         public virtual Iva Iva {get; set;}
